add func sort_grid_data to sys_utils.py, fix path: aims/get_ml_err.py
